fix: use new "mm args" cmd instead of Linux "ps" cmd for mm path
