+++ src/go/util/mm/mm.go
@@ -37,6 +37,7 @@ type MM interface {
 	GetNamespaceHosts(string) (Hosts, error)
 	Headnode() string
 	IsHeadnode(string) bool
+	GetMMArgs() map[string]string
 	GetVLANs(...Option) (map[string]int, error)
 
 	IsC2ClientActive(...C2Option) error
