+++ src/go/util/mm/package.go
@@ -108,6 +108,10 @@ func IsHeadnode(node string) bool {
 	return DefaultMM.IsHeadnode(node)
 }
 
+func GetMMArgs() map[string]string {
+	return DefaultMM.GetMMArgs()
+}
+
 func GetVLANs(opts ...Option) (map[string]int, error) {
 	return DefaultMM.GetVLANs(opts...)
 }
